feat: 重构时间计算逻辑并优化GPU调度
重构时间计算逻辑，使用统一的`nowMs`时间戳替代分散的性能计时。新增`WebGPUFrameEncoder`来优化GPU命令提交，减少编码器创建开销。调整插值和采样系统以支持无状态时间线时间，并修复播放控制器的反向播放问题。

- 将`performance.now()`替换为统一的`getNowMs()`工具函数
- 实现`WebGPUFrameEncoder`批量处理GPU命令
- 修改插值着色器直接使用`state.currentTime`
- 修复播放控制器在反向播放时的起始时间计算
- 更新测试用例验证新的时间计算逻辑

+++ packages/core/src/components/timeline/schema.ts
@@ -3,7 +3,7 @@ import { ComponentDef } from '../../plugin';
 export const TimelineComponent: ComponentDef = {
   schema: {
     tracks: 'object',
-    duration: 'float32',
+    duration: 'float64',
     loop: 'int32',
     repeat: 'int32',
     version: 'int32',
